improve subgraph retrival
Do subgraph retrival in a way that also walks and searches in the graph at the same time, so one does not need to open knowledge graph anymore (after calling retrieve_subgraphs.py). Also fix bug with one list to little, and save as pickle (with lists and dicts instead of str).